\name{phenotype}
\docType{data}
\alias{phenotype}
\title{Phenotypes of Genotyped Cattle}
\description{
 Phenotypes of cattle.
}
\usage{data(phenotype)}
\format{Data frame containing information on genotyped cattle. The columns contain the IDs of the individuals (\code{Indiv}), simulated breeding values (\code{BV}), simulated sexes (\code{Sex}), and genetic contributions from other breeds (\code{MC}).
}
\keyword{datasets}